fix: cast seg data to int32 before NiBabel image construction
NiBabel raises ValueError for int64 arrays unless an explicit header/dtype is provided. Cast case.seg.data to int32 (sufficient for all vertebra label values, which are at most a few hundred) before passing to nib.Nifti1Image.

+++ src/segqc/cli.py
@@ -163,7 +163,7 @@ def _handle_run(args: argparse.Namespace) -> int:
     import nibabel as nib  # noqa: PLC0415
 
     cfg = default_config()
-    seg_img = nib.Nifti1Image(case.seg.data, case.seg.affine)
+    seg_img = nib.Nifti1Image(case.seg.data.astype("int32"), case.seg.affine)
     check_result = check_empty(seg_img, cfg)
 
     # --- 4. Build Verdict from CheckResult ------------------------------------ #
